have the fields in error
When trying to create an account with empty fields

Now, when the fields is in error, we do not submit the page. an ajax
call is made to verify if fields are in error. If so, the field is in
red with a message explaining why, until the value is changed.