#5 have a screen to get forgotten password
When clicking on the link to get forgotten password, display the screen
to set an email. Click on continue to send a mail with the password and
be redirected to the authenticate screen.

If an error occured (For example if not account are found) we stay on
the same screen with an error message saying why